+++ Samples~/ImportTester/Code/ModelDisplayer.cs
@@ -1,10 +1,16 @@
+#nullable enable
+
 using UnityEngine;
+using UnityEngine.Events;
 
 namespace At.Ac.FhStp.Import3D.ImportTester
 {
     public class ModelDisplayer : MonoBehaviour
     {
         [SerializeField] private Transform modelParent = null!;
+
+        [SerializeField] private UnityEvent<Transform?> displayedModelChanged =
+            new UnityEvent<Transform?>();
 
         private int modelCount;
         private int? modelIndex;
@@ -16,6 +22,8 @@ namespace At.Ac.FhStp.Import3D.ImportTester
 
             var current = modelParent.GetChild(modelIndex.Value);
             current.gameObject.SetActive(true);
+
+            displayedModelChanged.Invoke(current);
         }
 
         public void OnModelsLoaded()
@@ -23,7 +31,10 @@ namespace At.Ac.FhStp.Import3D.ImportTester
             modelCount = modelParent.childCount;
             modelIndex = modelCount > 0 ? 0 : null;
 
-            UpdateDisplayedModel();
+            if (modelIndex != null)
+                UpdateDisplayedModel();
+            else
+                displayedModelChanged.Invoke(null);
         }
 
         public void OnModelSwitchInput(LinearDirection direction)